fix(providers): Move provider's logic outside main
